requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.7
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
Pillow==10.1.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
wrapt==1.16.0
